requirements.txt
altair==4.2.2
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
entrypoints==0.4
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.14.1
numpy==1.23.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.9.0
protobuf==3.20.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.28.0
rich==13.9.4
rpds-py==0.21.0
semver==3.0.2
six==1.16.0
smmap==5.0.1
streamlit==1.40.2
tenacity==9.0.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
tradingview-ta==3.2.10
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
validators==0.34.0
zipp==3.21.0
